but: non-legacy commit move
but uses the newly implemented graph-based move commit operations.
