+++ crates/but/src/command/commit/mod.rs
@@ -1,0 +1,2 @@
+pub mod file;
+pub mod r#move;
