+++ crates/but/src/utils/metrics.rs
@@ -189,7 +189,6 @@ impl Subcommands {
             Subcommands::Squash { .. } => Rub,
             #[cfg(feature = "legacy")]
             Subcommands::Merge { .. } => Merge,
-            #[cfg(feature = "legacy")]
             Subcommands::Move { .. } => Rub,
             #[cfg(feature = "legacy")]
             Subcommands::Pick { .. } => Pick,
